requirements.txt
# Quantum Image Retrieval System - Requirements
# Version 4.0.0 - Standardized Dependencies

# === Web Framework ===
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7

# === Azure Services ===
azure-cosmos==4.9.0
azure-identity==1.15.0
azure-core==1.29.5
azure-storage-blob==12.19.0

# === Configuration ===
python-dotenv==1.0.0

# === Deep Learning and Image Processing ===
torch==2.8.0
torchvision==0.18.0
Pillow==10.1.0
opencv-python==4.8.1.78

# === Scientific Computing ===
numpy==1.24.4
scipy==1.11.4

# === Quantum Computing ===
qiskit==1.0.2
qiskit-aer==0.13.3

# === Utilities ===
tqdm==4.66.1
requests==2.31.0
